added tests for day_04
