--- aoc2019/input/day_02_ex1
@@ -1,1 +1,0 @@
-1,9,10,3,2,3,11,0,99,30,40,50
