+++ aoc2019/input/day_05_ex_1
@@ -1,0 +1,1 @@
+3,12,6,12,15,1,13,14,13,4,13,99,-1,0,1,9
