+++ aoc2019/pyproject.toml
@@ -5,8 +5,12 @@ description = ""
 authors = ["Steven Powell <princeodd@gmail.com>"]
 
 [tool.poetry.scripts]
+d1p1 = "aoc2019.day_01:part1"
+d1p2 = "aoc2019.day_01:part2"
 d2p1 = "aoc2019.day_02:part1"
+d3p1 = "aoc2019.day_03:part1"
 d3p2 = "aoc2019.day_03:part2"
+d4p2 = "aoc2019.day_04:part2"
 d5p1 = "aoc2019.day_05:part1"
 d5p2 = "aoc2019.day_05:part2"
