+++ aoc2019/tests/test_day_02.py
@@ -6,7 +6,7 @@ from aoc2019.intcode import IntCode
 @pytest.mark.parametrize(
     ("input_file", "user_input", "expected_result"),
     [
-        ("input/day_02_ex1", 1, 3500),
+        ("input/day_02_ex_1", 1, 3500),
         ("input/day_02", 1, 5866714),
         ("input/day_02_p2", 0, 19690720)
     ]
