+++ aoc2019/tests/test_day_05.py
@@ -3,20 +3,22 @@ import pytest
 from aoc2019.intcode import IntCode
 
 
-def test_d5p1():
+def test_part1():
     intcode = IntCode()
     intcode.get_input("input/day_05")
     intcode.diagnostic_program(1)
     assert intcode.result == 16489636
 
-def test_d5_examples():
-    intcode = IntCode()
-    intcode.data = [3, 12, 6, 12, 15, 1, 13, 14, 13, 4, 13, 99, -1, 0, 1, 9]
-    intcode.diagnostic_program(0)
-    assert intcode.result == 0
 
-def test_d5p2():
+@pytest.mark.parametrize(
+    ("input_file", "user_input", "expected_result"),
+    [
+        ("input/day_05", 5, 9386583),
+        ("input/day_05_ex_1", 0, 0)
+    ]
+)
+def test_part2(input_file, user_input, expected_result):
     intcode = IntCode()
-    intcode.get_input("input/day_05")
-    intcode.diagnostic_program(5)
-    assert intcode.result == 9386583
+    intcode.get_input(input_file)
+    intcode.diagnostic_program(user_input)
+    assert intcode.result == expected_result
